Refactor S3Retention
