+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/runner/S3Distribution.java
@@ -20,20 +20,12 @@
 package app.coronawarn.server.services.distribution.runner;
 
 import app.coronawarn.server.services.distribution.assembly.component.OutputDirectoryProvider;
-import app.coronawarn.server.services.distribution.config.DistributionServiceConfig;
 import app.coronawarn.server.services.distribution.objectstore.ObjectStoreAccess;
-import app.coronawarn.server.services.distribution.objectstore.S3Object;
 import app.coronawarn.server.services.distribution.objectstore.S3Publisher;
 import io.minio.errors.MinioException;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.security.GeneralSecurityException;
-import java.time.LocalDate;
-import java.time.ZoneOffset;
-import java.time.format.DateTimeFormatter;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.ApplicationArguments;
@@ -54,20 +46,9 @@ public class S3Distribution implements ApplicationRunner {
 
   private final ObjectStoreAccess objectStoreAccess;
 
-  private final DistributionServiceConfig distributionServiceConfig;
-
-  /**
-   * Creates an S3Distribution object, which will upload and remove files from the S3-compatible storage.
-   *
-   * @param outputDirectoryProvider   the outputDirectoryProvider.
-   * @param objectStoreAccess         the objectStoreAccess for the S3Distribution.
-   * @param distributionServiceConfig the distributionServiceConfig, used to retrieve the retention days.
-   */
-  public S3Distribution(OutputDirectoryProvider outputDirectoryProvider, ObjectStoreAccess objectStoreAccess,
-      DistributionServiceConfig distributionServiceConfig) {
+  S3Distribution(OutputDirectoryProvider outputDirectoryProvider, ObjectStoreAccess objectStoreAccess) {
     this.outputDirectoryProvider = outputDirectoryProvider;
     this.objectStoreAccess = objectStoreAccess;
-    this.distributionServiceConfig = distributionServiceConfig;
   }
 
   @Override
@@ -80,45 +61,6 @@ public class S3Distribution implements ApplicationRunner {
       logger.info("Data pushed to CDN successfully.");
     } catch (UnsupportedOperationException | GeneralSecurityException | MinioException | IOException e) {
       logger.error("Distribution failed.", e);
-    }
-  }
-
-  /**
-   * Deletes all diagnosis-key files from S3 that are older than retentionDays.
-   *
-   * @param retentionDays the number of days, that files should be retained on S3.
-   */
-  public void applyRetentionPolicy(int retentionDays) throws MinioException, GeneralSecurityException, IOException {
-    List<S3Object> diagnosisKeysObjects = this.objectStoreAccess.getObjectsWithPrefix("version/v1/"
-        + distributionServiceConfig.getApi().getDiagnosisKeysPath() + "/"
-        + distributionServiceConfig.getApi().getCountryPath() + "/"
-        + distributionServiceConfig.getApi().getCountryGermany() + "/"
-        + distributionServiceConfig.getApi().getDatePath() + "/");
-    final String regex = ".*([0-9]{4}-[0-9]{2}-[0-9]{2}).*";
-    final Pattern pattern = Pattern.compile(regex);
-
-    final LocalDate cutOffDate = LocalDate.now(ZoneOffset.UTC).minusDays(retentionDays);
-
-    diagnosisKeysObjects.stream()
-        .filter(diagnosisKeysObject -> {
-          Matcher matcher = pattern.matcher(diagnosisKeysObject.getObjectName());
-          return matcher.matches() && LocalDate.parse(matcher.group(1), DateTimeFormatter.ISO_LOCAL_DATE)
-              .isBefore(cutOffDate);
-        })
-        .forEach(this::deleteDiagnosisKey);
-  }
-
-  /**
-   * Java stream do not support checked exceptions within streams. This helper method rethrows them as unchecked
-   * expressions, so they can be passed up to the Retention Policy.
-   *
-   * @param diagnosisKey the  diagnosis key, that should be deleted.
-   */
-  public void deleteDiagnosisKey(S3Object diagnosisKey) {
-    try {
-      objectStoreAccess.deleteObjectsWithPrefix(diagnosisKey.getObjectName());
-    } catch (Exception e) {
-      throw new RuntimeException(e);
     }
   }
 }
